Refactor file structures to add animations to every slide
