+++ client/components/Resume/DesktopView/DisplayBox/Slides/Skills/style.js
@@ -1,0 +1,16 @@
+module.exports.body = {
+  position: 'relative',
+  right: '-1000px',
+  height: '100%',
+  width: '100%',
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  justifyContent: 'space-evenly',
+  backgroundColor: 'rgb(250, 250, 250)',
+  WebkitTransition: '0.2s'
+};
+
+module.exports.body_loaded = {
+  right: '0px'
+};
